requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
anthropic>=0.25
elevenlabs>=1.9
pinecone-client>=5.0
pymongo>=4.10
motor>=3.3
structlog>=24.1
httpx>=0.27
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
pytest>=8.1
pytest-asyncio>=0.23
ruff>=0.3
